Refine filter and formula transforms.
