+++ src/js/components/pipelines/HoverField.jsx
@@ -9,9 +9,9 @@ var React = require('react'),
     bindChannel = require('../../actions/bindChannel'),
     Icon = require('../Icon'),
     FieldType = require('./FieldType'),
-    SortIcon = require('./transforms/Icons/Sort'),
-    FilterIcon = require('./transforms/Icons/Filter').connected,
-    FormulaIcon = require('./transforms/Icons/Formula').connected,
+    SortIcon = require('./transforms/SortIcon'),
+    FilterIcon = require('./transforms/FilterIcon').connected,
+    FormulaIcon = require('./transforms/FormulaIcon').connected,
     AggregateList = require('./AggregateList'),
     getInVis = require('../../util/immutable-utils').getInVis,
     assets = require('../../util/assets'),
